requirements/production.txt
bs4
newspaper3k
